RAB-686: Move ClientStorageInterface and add event listener in order to handle job execution error
