+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Domain/StorageClientInterface.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-namespace Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure\StorageClient;
+namespace Akeneo\Platform\Bundle\ImportExportBundle\Domain;
 
 interface StorageClientInterface
 {
