+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Infrastructure/StorageClient/StorageClientProviderInterface.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure\StorageClient;
 
 use Akeneo\Platform\Bundle\ImportExportBundle\Domain\Model\StorageInterface;
+use Akeneo\Platform\Bundle\ImportExportBundle\Domain\StorageClientInterface;
 
 interface StorageClientProviderInterface
 {
