+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Infrastructure/StorageClient/TransferFilesToStorage.php
@@ -10,14 +10,17 @@ declare(strict_types=1);
 namespace Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure\StorageClient;
 
 use Akeneo\Platform\Bundle\ImportExportBundle\Application\TransferFilesToStorage\FileToTransfer;
+use Akeneo\Platform\Bundle\ImportExportBundle\Domain\Event\FileCannotBeExported;
 use Akeneo\Platform\Bundle\ImportExportBundle\Domain\Model\StorageInterface;
 use Akeneo\Platform\Bundle\ImportExportBundle\Domain\TransferFilesToStorageInterface;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 
 final class TransferFilesToStorage implements TransferFilesToStorageInterface
 {
     public function __construct(
         private StorageClientProvider $storageClientProvider,
-        private TransferFile $transferFile
+        private TransferFile $transferFile,
+        private EventDispatcher $eventDispatcher
     ) {
     }
 
@@ -30,12 +33,16 @@ final class TransferFilesToStorage implements TransferFilesToStorageInterface
         foreach ($filesToTransfer as $fileToTransfer) {
             $sourceStorage = $this->storageClientProvider->getFromFileToTransfer($fileToTransfer);
 
-            $this->transferFile->transfer(
-                $sourceStorage,
-                $destinationStorage,
-                $fileToTransfer->getFileKey(),
-                $fileToTransfer->getOutputFileName()
-            );
+            try {
+                $this->transferFile->transfer(
+                    $sourceStorage,
+                    $destinationStorage,
+                    $fileToTransfer->getFileKey(),
+                    $fileToTransfer->getOutputFileName()
+                );
+            } catch (\Exception $exception) {
+                $this->eventDispatcher->dispatch(new FileCannotBeExported($exception->getMessage()));
+            }
         }
     }
 }
