+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Resources/config/event_subscribers.yml
@@ -4,5 +4,6 @@ services:
       - '@akeneo_batch.job.job_registry'
       - '@Akeneo\Platform\Bundle\ImportExportBundle\Application\TransferFilesToStorage\TransferFilesToStorageHandler'
       - '@feature_flags'
+      - '@event_dispatcher'
     tags:
       - { name: kernel.event_subscriber }
