+++ tests/back/Platform/Specification/Bundle/ImportExportBundle/Infrastructure/StorageClient/StorageClientProviderSpec.php
@@ -15,7 +15,7 @@ namespace Specification\Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure
 
 use Akeneo\Platform\Bundle\ImportExportBundle\Application\TransferFilesToStorage\FileToTransfer;
 use Akeneo\Platform\Bundle\ImportExportBundle\Domain\Model\StorageInterface;
-use Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure\StorageClient\StorageClientInterface;
+use Akeneo\Platform\Bundle\ImportExportBundle\Domain\StorageClientInterface;
 use Akeneo\Platform\Bundle\ImportExportBundle\Infrastructure\StorageClient\StorageClientProviderInterface;
 use Akeneo\Tool\Component\FileStorage\FilesystemProvider;
